Altered '/' chat commands to make wildcard more accurate
So we had noticed that the / commands available in chat would trigger if you posted the keyword in the middle of a message. 

More specifically, sending the message 'foo bar google.com/skipit foo' would skip the current track.

This is due to the if statements in sendMessage( ), which looked for ' indexOf !== -1 ' , which resolves to 'If [word] exists in message'. Changed it to === 0, which is what is returned by indexOf when the substring is at the beginning of the word. 

Hope its ok if we contribute!! :)