Fixed playlist link not directing to right page
Currently the playlist button acts as a link to whichever menu you were last in in the playlist browser. AFAIK it's supposed to take you to the room's queue instead, since the other two buttons link to the two other playlist menus. This fixes it so you now have a button for each of the three menus.