Ignore case when checking if user is @mentioned
Ignore lower/upper case when checking if a new message @mentions the user; this currently only affects whether a mention sound is played or not. This fixes #248. E.g. with this change, a message with "@example" will trigger a notification sound for a user "eXampLe".

(i also removed the global regexp flag, because it's not needed here, as we are only testing if the message contains *any* @mention for the current user)