improve readability of usernames in chat
This makes the "i" and "l" distinguishable in usernames and @mentions in the chat, without having to change the font (-family). It creates a fake bold effect by adding a non-blurry text-shadow, that will adapt to the current text colour (so it should work for script users with custom styles).

Only conflict I noticed was with how DubX styles the names of admins (it manually adds a gray text-shadow that overrides the one of this commit). But it looks nice anyways, so ¯\\\_(ツ)\_/¯

before:
![](https://i.imgur.com/nHv52KZ.png) ![](https://i.imgur.com/313xArV.png) ![](https://i.imgur.com/4SVxeKu.png)

after:
![](https://i.imgur.com/xYP8Taz.png) ![](https://i.imgur.com/ChFPVM3.png) ![](https://i.imgur.com/kkZvT7C.png)
as you can see, "i"s definitely have an i-dot with this commit, instead of looking like straight lines (like a lowercase L)

It does **not** improve the readability of emotes ( /me messages ) ![](https://i.imgur.com/gFoZxqr.png)